requirements.txt
requests>=2.31,<3
beautifulsoup4>=4.12,<5
tldextract>=5.1,<6
openpyxl>=3.1,<4
python-whois>=0.8,<1
playwright>=1.45,<2
Pillow>=10.3,<11
imagehash>=4.3,<5
pytesseract>=0.3.10,<0.4
aiohttp>=3.9,<4
